Fix ast.Schema inferred root types on Mutation & Subscription and `schema` definition
The GraphQL spec defines an inference rule where a valid Schema can omit the `schema` definition and some common root type's name will be inferred as the schema's root(s).

However, when a `schema` definition is explicitely provided, it should be respected and the inference rule should not apply.

Right now, the inference is performed as soon as the `ast.Schema.<RootType>` is `nil`. This causes downstream gqlgen models generator to refuse to generate this type because it uses a condition `schemaType == cfg.Schema.Subscription` to exclude the root type from being generated which ultimately breaks the code compilation since a a type is missing.

The fix applied here is to perform the inference rule only if no `schema` definition was provided.